code for generating curves shifted to new file for reusing purpose  , routing done using getx  instead of Navigator
