+++ lib/main.dart
@@ -3,6 +3,7 @@ import 'package:Liveasy/screens/new_login_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:Liveasy/screens/new_loading_screen.dart';
 import 'package:firebase_core/firebase_core.dart';
+import 'package:get/get.dart';
 
 
 // class FlashChat extends StatelessWidget {
@@ -69,7 +70,7 @@ class FlashChat extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return GetMaterialApp(
       theme: ThemeData(),
       initialRoute: NewLoginScreen.routeName,
       routes: {
