+++ pubspec.yaml
@@ -49,6 +49,7 @@ dependencies:
   flutter_spinkit: ^5.0.0
   firebase_core: ^1.2.0
   firebase_auth: ^1.2.0
+  get: ^4.1.4
 
 
 dev_dependencies:
